Create e2e dialogues task and dataset files
