+++ genienlp/tasks/__init__.py
@@ -27,4 +27,4 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from . import almond_task, generic_task, hf_task  # noqa
+from . import almond_task, dialogue_task, generic_task, hf_task  # noqa
